chore: update form component
